requirements.txt
numpy==1.19.5
streamlit==0.81.0
keras==2.4.3
Pillow==8.2.0
tensorflow==2.5.0
opencv-python==4.5.2.52